% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{path_params}
\alias{path_params}
\title{Deconstruct a path pattern}
\usage{
path_params(path)
}
\arguments{
\item{path}{The path pattern to parse}
}
\value{
A list with the elements \code{keys} containing the names of all the path
parameters and \code{glue} containing a glue ready version of the path
}
\description{
This function parses a path pattern and returns both the name of the
parameters and a version of the path formatted for glue string
interpolation.
}
\examples{
path_params("/users/:user/assets/*")
}
