% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Waysign.R
\name{Waysign}
\alias{Waysign}
\title{A simple, multi-purpose router}
\description{
The \code{Waysign} class implements a high-performance, multipurpose router build
on top of the \href{https://github.com/viz-rs/path-tree}{path-tree} library. A
router associates filepath-like patterns with a piece of data for latter
retrieval. Often that data is a function and the path to be matched against
the pattern comes from a URL, but it can be anything, adapting to the need
of the user.
}
\details{
The path pattern supported by Waysign mirrors that of path-tree and while the
full documentation can be found there, it will be briefly explained here.

A path pattern consist of zero, one, or more elements separated by \code{/}
(always started by \code{/}). Each element can either be a literal or one of the
following variable types:
\itemize{
\item \verb{:name} matches a single path piece
\item \verb{:name?} matches an optional path piece
\item \verb{:name+} or \code{+} matches one or more path pieces
\item \verb{:name*} or \code{*} matches zero or more path pieces
}

A variable don't have to consume a full path element. E.g. you could have a
path pattern like this: \verb{/date/:day-:month-:year} which would match to paths
such as \verb{/date/24-12-2025}
}
\examples{
# Adapted from path-tree docs
router <- Waysign$new()

router$add_path("/", 1)
router$add_path("/login", 2)
router$add_path("/signup", 3)
router$add_path("/settings", 4)
router$add_path("/settings/:page", 5)
router$add_path("/:user", 6)
router$add_path("/:user/:repo", 7)
router$add_path("/public/:any*", 8)
router$add_path("/:org/:repo/releases/download/:tag/:filename.:ext", 9)
router$add_path("/:org/:repo/tags/:day-:month-:year", 10)
router$add_path("/:org/:repo/actions/:name\\\\::verb", 11)
router$add_path("/:org/:repo/:page", 12)
router$add_path("/:org/:repo/*", 13)
router$add_path("/api/+", 14)

router$find_object("/")
router$find_object("/login")
router$find_object("/settings/admin")
router$find_object("/viz-rs")
router$find_object("/viz-rs/path-tree")
router$find_object("/rust-lang/rust-analyzer/releases/download/2022-09-12/rust-analyzer-aarch64-apple-darwin.gz")
router$find_object("/rust-lang/rust-analyzer/tags/2022-09-12")
router$find_object("/rust-lang/rust-analyzer/actions/ci:bench")
router$find_object("/rust-lang/rust-analyzer/stargazers")
router$find_object("/rust-lang/rust-analyzer/stargazers/404")
router$find_object("/public/js/main.js")
router$find_object("/api/v1")

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Waysign-new}{\code{Waysign$new()}}
\item \href{#method-Waysign-format}{\code{Waysign$format()}}
\item \href{#method-Waysign-add_path}{\code{Waysign$add_path()}}
\item \href{#method-Waysign-find_object}{\code{Waysign$find_object()}}
\item \href{#method-Waysign-remove_path}{\code{Waysign$remove_path()}}
\item \href{#method-Waysign-clone}{\code{Waysign$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Waysign-new"></a>}}
\if{latex}{\out{\hypertarget{method-Waysign-new}{}}}
\subsection{Method \code{new()}}{
Create a new Waysign router
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waysign$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Waysign-format"></a>}}
\if{latex}{\out{\hypertarget{method-Waysign-format}{}}}
\subsection{Method \code{format()}}{
Pretty printing of the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waysign$format(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Waysign-add_path"></a>}}
\if{latex}{\out{\hypertarget{method-Waysign-add_path}{}}}
\subsection{Method \code{add_path()}}{
Add a new path to the router. See \emph{Details} for allowed path syntax
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waysign$add_path(path, object)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{A string giving the path to add}

\item{\code{object}}{An R object to be associated with the path.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Waysign-find_object"></a>}}
\if{latex}{\out{\hypertarget{method-Waysign-find_object}{}}}
\subsection{Method \code{find_object()}}{
Search for a path in the router
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waysign$find_object(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The path to search for}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
If no matching path is found then \code{NULL}, otherwise a list with
the elements \code{path} giving the path pattern that was matched, \code{object}
giving the object associated with the path, \code{keys} giving the name of the
path parameters in the path element, and \code{values} giving the values of
the keys based on the match.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Waysign-remove_path"></a>}}
\if{latex}{\out{\hypertarget{method-Waysign-remove_path}{}}}
\subsection{Method \code{remove_path()}}{
Remove a path from the router. Due to the underlying implementation this
causes a complete rebuild of the router
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waysign$remove_path(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{A string giving the path to remove}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Waysign-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Waysign-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waysign$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
