% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waysign-package.R
\docType{package}
\name{waysign-package}
\alias{waysign}
\alias{waysign-package}
\title{waysign: Multi-Purpose and High-Performance Routing}
\description{
provides routing based on the 'path-tree' Rust crate. The routing is general purpose in the sense that any type of R object can be associated with a path, not just a handler function.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomasp85/waysign}
  \item Report bugs at \url{https://github.com/thomasp85/waysign/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

}
\keyword{internal}
